Minor bug fixes, malloc casts, and other mods to make the damn thing link against abiword. Next step is to place hooks in the code to get the character data out of a word file, next ver or two will have this, hold tough.
git-svn-id: svn+ssh://svn.abisource.com/svnroot/wv/trunk@20247 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ table.c
@@ -384,7 +384,7 @@ int do_tablelooks(pap *apap)
 
 void output_tablebg(pap *apap)
 	{
-	int shademodifier = 100;
+	float shademodifier = 100;
 	int foreground;
 
 	if (apap->ourtap.tlp.itl)
